added some handy compiler flags and readjusted the code according to the warnings with dynamic memory allocation (calloc)

+++ password_generaor/generator.h
@@ -1,6 +1,6 @@
 #ifndef GENERATOR_H
 #define GENERATOR_H
 
-void generator(int passwd_len);
+void generator(unsigned long passwd_len);
 
 #endif // GENERATOR_H
